move merge related tests to merge package and added observation undo merge tests
